+++ src/main/scala/beam/agentsim/agents/modalbehaviors/ChoosesMode.scala
@@ -37,7 +37,7 @@ trait ChoosesMode {
   val dummyRHVehicle =
     StreetVehicle(
       Id.create("dummyRH", classOf[Vehicle]),
-      SpaceTime(0.0, 0.0, 0l),
+      SpaceTime(0, 0, 0),
       CAR,
       asDriver = false
     )
@@ -62,7 +62,7 @@ trait ChoosesMode {
     case Event(MobilityStatusResponse(streetVehicles), choosesModeData: ChoosesModeData) =>
       val bodyStreetVehicle = StreetVehicle(
         bodyId,
-        SpaceTime(currentActivity(choosesModeData.personData).getCoord, _currentTick.get.toLong),
+        SpaceTime(currentActivity(choosesModeData.personData).getCoord, _currentTick.get),
         WALK,
         asDriver = true
       )
@@ -120,7 +120,7 @@ trait ChoosesMode {
 
       def makeRideHailTransitRoutingRequest(bodyStreetVehicle: StreetVehicle): Option[Int] = {
         //TODO make ride hail wait buffer config param
-        val startWithWaitBuffer = 600 + departTime.atTime.toLong
+        val startWithWaitBuffer = 600 + departTime.atTime
         val currentSpaceTime =
           SpaceTime(currentActivity(choosesModeData.personData).getCoord, startWithWaitBuffer)
         val theRequest = RoutingRequest(
@@ -201,7 +201,7 @@ trait ChoosesMode {
                   val leg = BeamLeg(
                     departTime.atTime,
                     mode,
-                    l.getTravelTime.toLong,
+                    l.getTravelTime.toInt,
                     BeamPath(linkIds, None, SpaceTime.zero, SpaceTime.zero, r.getDistance)
                   )
                   router ! EmbodyWithCurrentTravelTime(leg, vehicle.id)
@@ -399,7 +399,7 @@ trait ChoosesMode {
           leg =>
             leg.copy(
               leg.beamLeg
-                .updateStartTime(startTimeAdjustment - startTimeBufferForWaiting.longValue())
+                .updateStartTime(startTimeAdjustment - startTimeBufferForWaiting.toInt)
           )
         ) ++ driveTransitTrip.legs.tail
         val fullTrip = if (rideHail2TransitEgressResult.error.isEmpty) {
