+++ src/main/scala/beam/router/BeamRouter.scala
@@ -151,7 +151,7 @@ class BeamRouter(
           )
           val transitDriver =
             context.actorOf(transitDriverAgentProps, transitDriverId.toString)
-          scheduler ! ScheduleTrigger(InitializeTrigger(0.0), transitDriver)
+          scheduler ! ScheduleTrigger(InitializeTrigger(0), transitDriver)
 
         case _ =>
           log.error(mode + " is not supported yet")
@@ -186,7 +186,7 @@ class BeamRouter(
                     transportNetwork.streetLayer.edgeStore.getCursor(edges.head)
                   val endEdge =
                     transportNetwork.streetLayer.edgeStore.getCursor(edges.last)
-                  (departureTime: Long, _: Int, vehicleId: Id[Vehicle]) =>
+                  (departureTime: Int, _: Int, vehicleId: Id[Vehicle]) =>
                     BeamPath(
                       edges.map(_.intValue()).toVector,
                       Option(TransitStopsInfo(fromStop, vehicleId, toStop)),
@@ -208,7 +208,7 @@ class BeamRouter(
                     .getCursor(edgeIds.head)
                   val endEdge = transportNetwork.streetLayer.edgeStore
                     .getCursor(edgeIds.last)
-                  (departureTime: Long, duration: Int, vehicleId: Id[Vehicle]) =>
+                  (departureTime: Int, duration: Int, vehicleId: Id[Vehicle]) =>
                     BeamPath(
                       edgeIds,
                       Option(TransitStopsInfo(fromStop, vehicleId, toStop)),
@@ -240,7 +240,7 @@ class BeamRouter(
                 transportNetwork.streetLayer.edgeStore.getCursor(edgeIds.head)
               val endEdge =
                 transportNetwork.streetLayer.edgeStore.getCursor(edgeIds.last)
-              (departureTime: Long, duration: Int, vehicleId: Id[Vehicle]) =>
+              (departureTime: Int, duration: Int, vehicleId: Id[Vehicle]) =>
                 BeamPath(
                   edgeIds,
                   Option(TransitStopsInfo(fromStop, vehicleId, toStop)),
@@ -278,10 +278,10 @@ class BeamRouter(
             case Array((departureTimeFrom, from), (_, to)) =>
               val duration = tripSchedule.arrivals(to) - departureTimeFrom
               legs += BeamLeg(
-                departureTimeFrom.toLong,
+                departureTimeFrom,
                 mode,
                 duration,
-                transitPaths(from)(departureTimeFrom.toLong, duration, tripVehId)
+                transitPaths(from)(departureTimeFrom, duration, tripVehId)
               )
           }
           (tripVehId, (route, legs))
